# fix: correctly parse documentation comments