Fixes #258.