@JonathanStarup 2021-12-14
<details>
 <summary>Throughput Benchmarks</summary>
**This PR:**
```
Throughput (best): 10.717 lines/sec (with 8 threads.)

  min:  3.063, max: 10.717, avg:  8.706, median: 10.024

  The highest throughput was in iteration: 8 (out of 14).
  The ratio between the best and worst iteration was: 3,5x.

Finished 14 iterations on 58.966 lines of code in 106 seconds.
```
**Master:**
```
====================== Flix Compiler Throughput ======================

Throughput (best): 10.721 lines/sec (with 8 threads.)

  min:  3.047, max: 10.721, avg:  8.932, median:  9.665

  The highest throughput was in iteration: 10 (out of 14).
  The ratio between the best and worst iteration was: 3,5x.

Finished 14 iterations on 58.966 lines of code in 103 seconds.
```
</details>