@JonathanStarup 2021-12-14
doc comments are seperated by anything other than space and tab
```
/// This is 

/// not a single doc comment

/// this is
     /// a single
  /// doc comment
```
And a defs/classes/etc. use the last defined doc comment, so this is valid
```
/// doc comment

// something
/* somehitng */
def f(): Unit = ()
```