@magnus-madsen 2021-12-14
Could you try `--Xbenchmark-phases` too?