@JonathanStarup 2021-12-14
# Phase Benchmarks
## This PR:
Best runtime per phase (low to high):
  Deriver                          1,1 ms (00,0%)
  Reader                           1,7 ms (00,0%)
  Instances                        3,7 ms (00,1%)
  Terminator                       6,6 ms (00,1%)
  Safety                           8,7 ms (00,2%)
  Tailrec                          9,7 ms (00,2%)
  VarNumbering                    17,6 ms (00,3%)
  Lowering                        18,6 ms (00,4%)
  Simplifier                      21,6 ms (00,4%)
  Eraser                          21,7 ms (00,4%)
  Optimizer                       22,4 ms (00,4%)
  Stratifier                      24,4 ms (00,5%)
  LambdaLift                      25,1 ms (00,5%)
  Kinder                          30,2 ms (00,6%)
  Weeder                          32,5 ms (00,6%)
  PatternExhaustiveness           46,7 ms (00,9%)
  ClosureConv                     56,4 ms (01,1%)
  TreeShaker                      78,4 ms (01,5%)
  Namer                           91,3 ms (01,8%)
  Redundancy                      98,1 ms (01,9%)
  Finalize                       125,1 ms (02,5%)
  Resolver                       269,9 ms (05,3%)
  Parser                         272,2 ms (05,3%)
  Monomorph                      742,3 ms (14,5%)
  JvmBackend                     1492,5 ms (29,2%)
  Typer                          1586,2 ms (31,1%)

Finished 14 iterations on 58.705 lines of code in 102 seconds.

## Master:
Best runtime per phase (low to high):
  Deriver                          1,0 ms (00,0%)
  Reader                           1,8 ms (00,0%)
  Instances                        3,3 ms (00,1%)
  Terminator                       6,2 ms (00,1%)
  Safety                           8,4 ms (00,2%)
  Tailrec                          9,6 ms (00,2%)
  VarNumbering                    16,5 ms (00,3%)
  Lowering                        20,6 ms (00,4%)
  Simplifier                      20,9 ms (00,4%)
  Stratifier                      21,9 ms (00,4%)
  Optimizer                       22,2 ms (00,4%)
  Eraser                          22,3 ms (00,4%)
  LambdaLift                      25,7 ms (00,5%)
  Weeder                          26,2 ms (00,5%)
  Kinder                          30,3 ms (00,6%)
  PatternExhaustiveness           46,6 ms (00,9%)
  ClosureConv                     54,8 ms (01,1%)
  TreeShaker                      72,1 ms (01,4%)
  Namer                           85,6 ms (01,7%)
  Redundancy                      93,5 ms (01,8%)
  Finalize                       123,5 ms (02,4%)
  Resolver                       239,7 ms (04,7%)
  Parser                         354,8 ms (07,0%)
  Monomorph                      713,0 ms (14,1%)
  JvmBackend                     1450,0 ms (28,6%)
  Typer                          1592,1 ms (31,4%)

Finished 14 iterations on 58.705 lines of code in 97 seconds.