@magnus-madsen 2021-12-15
Ill check this out local and see how it interacts with `api.flix.dev`. One thing that is missing:

In `Weeder` we should:

- [ ] Strip empty *beginning* lines.
- [ ] Strip empty *ending* lines.
- [ ] Trim all lines (i.e. remove whitespace before and after the line).

See `visitDoc`