@magnus-madsen 2021-12-15
When I compile `JonathanStarup:document-parsing` locally and run `--doc` I don't see any documentation (`"doc"`) at all. I guess something is broken?