@JonathanStarup 2021-12-15
Parsing is painful... The error was that my copied files had `\n` instead of `\r\n` and that `\r\n` is parsed as two new lines so all doc comments where split into their lines where only the last is considered...