@magnus-madsen 2021-12-15
> > In the file `target/api/api.json`.
> 
> I can't run `--doc` for some reason. `java.lang.RuntimeException: Unable to write to path '.\target\api\api.json'.` even with admin

Maybe you need to create the folder `target/api`.