@magnus-madsen 2021-12-15
I think it would be a one-line fix in `Documentor`.