@JonathanStarup 2021-12-15
<details>
<summary>New Benchmarks</summary>
--THIS PR--
Best runtime per phase (low to high):
  Deriver                          1,1 ms (00,0%)
  Reader                           1,8 ms (00,0%)
  Instances                        3,1 ms (00,1%)
  Terminator                       9,4 ms (00,2%)
  Tailrec                         10,0 ms (00,2%)
  Safety                          10,8 ms (00,2%)
  VarNumbering                    17,9 ms (00,3%)
  Lowering                        21,7 ms (00,4%)
  Optimizer                       21,9 ms (00,4%)
  Eraser                          22,0 ms (00,4%)
  Simplifier                      22,0 ms (00,4%)
  PatternExhaustiveness           22,7 ms (00,4%)
  Stratifier                      25,8 ms (00,5%)
  LambdaLift                      26,6 ms (00,5%)
  Weeder                          34,7 ms (00,7%)
  Kinder                          36,1 ms (00,7%)
  ClosureConv                     53,8 ms (01,0%)
  TreeShaker                      73,8 ms (01,4%)
  Namer                           82,7 ms (01,6%)
  Redundancy                      90,9 ms (01,8%)
  Finalize                       121,5 ms (02,4%)
  Resolver                       265,9 ms (05,2%)
  Parser                         295,0 ms (05,7%)
  Monomorph                      709,2 ms (13,8%)
  JvmBackend                     1555,6 ms (30,3%)
  Typer                          1606,1 ms (31,2%)

Finished 14 iterations on 58.384 lines of code in 101 seconds.
--MASTER--
</details>