@magnus-madsen 2021-12-15
> Okay, RFR. when I run `--doc` theres is only 80/1400 ish empty doc comments

Nice. I will try it out at home later with `api.flix.dev`.