@magnus-madsen 2021-12-16
It seems to work for classes and enums (👍 ) but for signatures (defs inside classes) I don't get any docs.